requirements.txt
gunicorn==20.1.0
rq==1.10.1
label-studio-ml>=1.0.8
matplotlib>=3.2.2
numpy>=1.18.5
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1

--extra-index-url https://download.pytorch.org/whl/cpu
torch
torch>=1.7.0
torchvision>=0.8.1
tqdm>=4.64.0
tensorboard>=2.4.1
pandas>=1.1.4
seaborn>=0.11.0
ipython  # interactive notebook
psutil  # system utilization
thop>=0.1.1  # FLOPs computation
omegaconf>=2.3.0
ultralytics
opencv-python-headless>=4.6.0
hydra-core
pyrootutils